Switch CLI command names
All the existing commands have been replaced with new commands which
read a bit better and leave better room for more commands in the future.

The following changes have been made:
- `reshape start` -> `reshape migration start`
- `reshape complete` -> `reshape migration complete`
- `reshape abort` -> `reshape migration abort`
- `reshape generate-schema-query` -> `reshape schema-query`

The old commands are still available but are marked as deprecated in the
help output and will be removed in the future. The README will be
updated later.

Closes #10